requirements.txt
google-adk>=0.0.1
google-cloud-aiplatform[adk,agent-engines]>=1.88.0
llama-index>=0.12
google-cloud-storage